requirements.txt
sqlalchemy
pymysql
python-dotenv
plotly